+++ lib/src/pigeons/point_annotation_messenger.dart
@@ -422,6 +422,7 @@ class PointAnnotation {
   /// Controls the transition progress between the image variants of icon-image. Zero means the first variant is used, one is the second, and in between they are blended together. . Both images should be the same size and have the same type (either raster or vector).
   /// Default value: 0. Value range: [0, 1]
   /// Deprecated: Use `PointAnnotationManager.iconImageCrossFade` instead.
+  @Deprecated("Use [PointAnnotationManager.iconImageCrossFade] instead.")
   double? iconImageCrossFade;
 
   /// The opacity at which the icon will be drawn in case of being depth occluded. Absent value means full occlusion against terrain only.
@@ -769,6 +770,7 @@ class PointAnnotationOptions {
   /// Controls the transition progress between the image variants of icon-image. Zero means the first variant is used, one is the second, and in between they are blended together. . Both images should be the same size and have the same type (either raster or vector).
   /// Default value: 0. Value range: [0, 1]
   /// Deprecated: Use `PointAnnotationManager.iconImageCrossFade` instead.
+  @Deprecated("Use [PointAnnotationManager.iconImageCrossFade] instead.")
   double? iconImageCrossFade;
 
   /// The opacity at which the icon will be drawn in case of being depth occluded. Absent value means full occlusion against terrain only.